refactor: move login exceptions and gateway client check to config
- Extract login exceptions into fb/login/exception.h
- Remove unused login gateway module
- Read gateway client version/nation from config instead of hardcoded util
