+++ server/login/src/handler/protocol/create_account.cpp
@@ -1,4 +1,5 @@
 #include <fb/login/handler/protocol/create_account.h>
+#include <fb/login/exception.h>
 #include <fb/model/model.h>
 #include <fb/encoding.h>
 #include <json/json.h>
